agents/orchestrator/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
httpx>=0.27.0
asyncpg>=0.30.0
pyjwt[crypto]>=2.8.0
pydantic>=2.0.0
anthropic>=0.45.0
google-cloud-aiplatform>=1.70.0
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0
opentelemetry-instrumentation-fastapi>=0.46b0
opentelemetry-instrumentation-httpx>=0.46b0
opentelemetry-exporter-gcp-trace>=1.7.0
